Refactor autoconfiguration, adjust cors for webflux
