+++ jwt-server/spring/jwt-spring-auth0-common/src/main/java/org/entur/jwt/spring/auth0/properties/Auth0AuthorityMapperProperties.java
@@ -1,4 +1,4 @@
-package org.entur.jwt.spring.properties;
+package org.entur.jwt.spring.auth0.properties;
 
 public class Auth0AuthorityMapperProperties {
