+++ jwt-server/spring/jwt-spring-auth0-common/src/main/java/org/entur/jwt/spring/auth0/properties/Auth0Properties.java
@@ -1,4 +1,4 @@
-package org.entur.jwt.spring.properties;
+package org.entur.jwt.spring.auth0.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
