+++ jwt-server/spring/jwt-spring-auth0-webflux/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.entur.jwt.spring.reactive.auth0.Auth0SecurityAutoConfigurationWebFlux
+org.entur.jwt.spring.auth0.webflux.Auth0WebFluxAutoConfiguration
