+++ jwt-server/spring/jwt-spring-common/src/main/java/org/entur/jwt/spring/auth0/properties/PermitAll.java
@@ -1,4 +1,4 @@
-package org.entur.jwt.spring.properties;
+package org.entur.jwt.spring.auth0.properties;
 
 /**
  *
