+++ jwt-server/spring/jwt-spring-web/src/main/java/org/entur/jwt/spring/JwtWebAutoConfiguration.java
@@ -1,62 +1,64 @@
 package org.entur.jwt.spring;
 
-import org.entur.jwt.spring.filter.JwtAuthenticationExceptionHandler;
-import org.entur.jwt.spring.filter.JwtAuthenticationManager;
+import org.entur.jwt.spring.auth0.properties.AuthorizationProperties;
+import org.entur.jwt.spring.auth0.properties.CorsProperties;
+import org.entur.jwt.spring.auth0.properties.PermitAll;
+import org.entur.jwt.spring.auth0.properties.SecurityProperties;
+import org.entur.jwt.spring.filter.JwtAuthenticationExceptionAdvice;
+import org.entur.jwt.spring.filter.JwtAuthenticationFilter;
 import org.entur.jwt.spring.filter.JwtAuthorityMapper;
 import org.entur.jwt.spring.filter.JwtDetailsMapper;
 import org.entur.jwt.spring.filter.JwtPrincipalMapper;
-import org.entur.jwt.spring.filter.JwtServerAuthenticationConverter;
 import org.entur.jwt.spring.filter.log.JwtMappedDiagnosticContextMapper;
 import org.entur.jwt.spring.filter.resolver.JwtArgumentResolver;
-import org.entur.jwt.spring.properties.AuthorizationProperties;
-import org.entur.jwt.spring.properties.CorsProperties;
-import org.entur.jwt.spring.properties.PermitAll;
-import org.entur.jwt.spring.properties.SecurityProperties;
 import org.entur.jwt.verifier.JwtClaimExtractor;
 import org.entur.jwt.verifier.JwtVerifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.ReactiveAuthenticationManager;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.reactive.CorsConfigurationSource;
-import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource;
-import org.springframework.web.reactive.config.WebFluxConfigurer;
-import org.springframework.web.reactive.result.method.annotation.ArgumentResolverConfigurer;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.HandlerExceptionResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
 @ConditionalOnProperty(name = {"entur.jwt.enabled"}, havingValue = "true")
-public class JwtAutoConfigurationWebFlux extends JwtAutoConfiguration {
+@AutoConfigureAfter(value = JwtAutoConfiguration.class)
+public class JwtWebAutoConfiguration {
 
-    private static final Logger log = LoggerFactory.getLogger(JwtAutoConfigurationWebFlux.class);
+    private static final Logger log = LoggerFactory.getLogger(JwtWebAutoConfiguration.class);
 
     @Configuration
-    public static class DefaultEnturWebFluxConfigurer implements WebFluxConfigurer {
+    public static class DefaultEnturWebMvcConfigurer implements WebMvcConfigurer {
 
         private final JwtArgumentResolver resolver;
 
         @Autowired
-        public DefaultEnturWebFluxConfigurer(JwtArgumentResolver resolver) {
+        public DefaultEnturWebMvcConfigurer(JwtArgumentResolver resolver) {
             this.resolver = resolver;
         }
 
         @Override
-        public void configureArgumentResolvers(ArgumentResolverConfigurer configurer) {
-            configurer.addCustomResolver(resolver);
+        public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+            resolvers.add(resolver);
         }
     }
 
     @Bean
-    @ConditionalOnMissingBean(JwtServerAuthenticationConverter.class)
-    public <T> JwtServerAuthenticationConverter<T> auth(SecurityProperties properties, JwtVerifier<T> verifier, @Autowired(required = false) JwtMappedDiagnosticContextMapper<T> mdcMapper, JwtAuthorityMapper<T> authorityMapper,
-                                                        JwtClaimExtractor<T> extractor, JwtPrincipalMapper jwtPrincipalMapper, JwtDetailsMapper jwtDetailsMapper) {
+    @ConditionalOnMissingBean(JwtAuthenticationFilter.class)
+    public <T> JwtAuthenticationFilter<T> auth(SecurityProperties properties, JwtVerifier<T> verifier, @Autowired(required = false) JwtMappedDiagnosticContextMapper<T> mdcMapper, JwtAuthorityMapper<T> authorityMapper,
+                                               JwtClaimExtractor<T> extractor, @Lazy HandlerExceptionResolver handlerExceptionResolver, JwtPrincipalMapper jwtPrincipalMapper, JwtDetailsMapper jwtDetailsMapper) {
         AuthorizationProperties authorizationProperties = properties.getAuthorization();
 
         PermitAll permitAll = authorizationProperties.getPermitAll();
@@ -68,34 +70,13 @@ public class JwtAutoConfigurationWebFlux extends JwtAutoConfiguration {
         } else {
             log.info("Authentication with Json Web Token is optional");
         }
-        return new JwtServerAuthenticationConverter<>(verifier, authorityMapper, mdcMapper, extractor, tokenMustBePresent, jwtDetailsMapper, jwtPrincipalMapper);
+        return new JwtAuthenticationFilter<>(verifier, tokenMustBePresent, authorityMapper, mdcMapper, extractor, handlerExceptionResolver, jwtPrincipalMapper, jwtDetailsMapper);
     }
 
     @Bean
-    @ConditionalOnMissingBean(JwtAuthenticationExceptionHandler.class) // allow for customization
-    public JwtAuthenticationExceptionHandler advice() {
-        return new JwtAuthenticationExceptionHandler();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean(ReactiveAuthenticationManager.class)
-    @ConditionalOnProperty(name = "entur.jwt.enabled", havingValue = "true")
-    public ReactiveAuthenticationManager reactiveAuthenticationManager() {
-        return new JwtAuthenticationManager();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean(CustomAuthenticationFailureHandler.class)
-    @ConditionalOnProperty(name = "entur.jwt.enabled", havingValue = "true")
-    public CustomAuthenticationFailureHandler customAuthenticationFailureHandler() {
-        return new CustomAuthenticationFailureHandler();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean(CustomServerAuthenticationEntryPoint.class)
-    @ConditionalOnProperty(name = "entur.jwt.enabled", havingValue = "true")
-    public CustomServerAuthenticationEntryPoint customServerAuthenticationEntryPoint() {
-        return new CustomServerAuthenticationEntryPoint();
+    @ConditionalOnMissingBean(JwtAuthenticationExceptionAdvice.class) // allow for customization
+    public JwtAuthenticationExceptionAdvice advice() {
+        return new JwtAuthenticationExceptionAdvice();
     }
 
     @Bean("corsConfigurationSource")
