+++ jwt-server/spring/jwt-spring-web/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.entur.jwt.spring.JwtAutoConfigurationWeb,org.entur.jwt.spring.JwtWebSecurityConfigurerAdapterAutoConfiguration
+org.entur.jwt.spring.JwtWebAutoConfiguration,org.entur.jwt.spring.JwtWebSecurityConfigurerAdapterAutoConfiguration
