+++ jwt-server/spring/jwt-spring-webflux/src/main/java/org/entur/jwt/spring/JwtWebSecurityAutoConfiguration.java
@@ -1,14 +1,18 @@
 package org.entur.jwt.spring;
 
+import org.entur.jwt.spring.auth0.properties.SecurityProperties;
 import org.entur.jwt.spring.config.AuthorizationWebSecurityConfig;
 import org.entur.jwt.spring.config.JwtFilterWebSecurityConfig;
 import org.entur.jwt.spring.filter.JwtServerAuthenticationConverter;
-import org.entur.jwt.spring.properties.SecurityProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
-import org.springframework.boot.autoconfigure.condition.*;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -31,7 +35,7 @@ import org.springframework.security.web.server.authentication.ServerAuthenticati
 @ConditionalOnClass(WebSecurityConfigurerAdapter.class)
 @ConditionalOnExpression("${entur.authorization.enabled:true} || ${entur.jwt.enabled:true}")
 @EnableConfigurationProperties({SecurityProperties.class})
-@AutoConfigureAfter(JwtAutoConfigurationWebFlux.class)
+@AutoConfigureAfter(JwtWebFluxAutoConfiguration.class)
 public class JwtWebSecurityAutoConfiguration {
 
     private static final Logger log = LoggerFactory.getLogger(JwtWebSecurityAutoConfiguration.class);
