+++ jwt-server/spring/jwt-spring-webflux/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.entur.jwt.spring.JwtAutoConfigurationWebFlux,org.entur.jwt.spring.JwtWebSecurityAutoConfiguration
+org.entur.jwt.spring.JwtWebFluxAutoConfiguration,org.entur.jwt.spring.JwtWebSecurityAutoConfiguration
